force lower case for package name
To be able to use `iceoryx`s native applications with `ros 2` based applications the package name defined from service and event name must be valid for `ros 2` as package name. Forcing lowercase in the package name only partially fixes this issue. As far as I know there are changes planned regarding types in `iceoryx`, so I would propose this fix, till we can focus on the new type handling.

Signed-off-by: Johannes Jeising <Johannes.Jeising@de.bosch.com>